feat(chore): Extraction de GristClient
